requirements.txt
# web server with Flask and Bootstrap4
pyFlaskBootstrap4>=0.2.8
# mySQL support
# https://pypi.org/project/PyMySQL/
PyMySQL
# Mediawiki and SMW API
# https://pypi.org/project/py-3rdparty-mediawiki/
py-3rdparty-mediawiki>=0.1.23
# https://phabricator.wikimedia.org/T171886
# https://pypi.org/project/pywikibot/
pywikibot>=5.3.0
#https://pypi.org/project/pycrypto/
pycrypto
#https://pypi.org/project/pydevd/
pydevd
# https://pypi.org/project/pylodstorage/
pylodstorage>=0.0.21
# parser
lxml
